requirements.txt
Django==5.0.2
python-decouple==3.8
mysqlclient==2.2.4
djangorestframework-simplejwt==5.3.1
djangorestframework==3.14.0
django-password-reset==2.0
pillow==10.2.0